Separated BatchFile and plotting functions
